move container data tagging to utils to resuse in generate recommendations
